Use the list of commit messages as the pull request description
When you create a pull request from the web interface, commit messages
go to the description field automatically. This commit does the same for
the CLI.

It also caches the result of the `git log` command as it is now used
twice (for title and description).

Also, some code has been made more rubyish :)